+++ src/app/app.component.ts
@@ -1,10 +1,15 @@
-import { Component } from '@angular/core';
+import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
-  title = 'market';
+export class AppComponent implements AfterViewInit {
+  @ViewChild('appTitle') appTitle!: ElementRef<HTMLHeadingElement>;
+  title = 'Market';
+
+  ngAfterViewInit() {
+    this.appTitle.nativeElement.innerText = this.title;
+  }
 }