+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,0 +1,30 @@
+import {Component, OnInit} from '@angular/core';
+import {ProductService} from "../../services/product.service";
+import {ProductModel} from "../../models/product.model";
+import {CartService} from "../../../cart/services/cart.service";
+
+@Component({
+  selector: 'app-products-list',
+  templateUrl: './product-list.component.html',
+  styleUrls: ['./product-list.component.scss']
+})
+export class ProductListComponent implements OnInit {
+
+  products!: ProductModel[];
+
+  constructor(public productService: ProductService,
+              private cartService: CartService) {
+  }
+
+  ngOnInit(): void {
+    this.products = this.productService.getProducts();
+  }
+
+  onAddToCart(product: ProductModel): void {
+    try {
+      this.cartService.addToCart(product)
+    } catch (error) {
+      confirm('You already have it in your cart')
+    }
+  }
+}