+++ changelog.md
@@ -7,3 +7,15 @@
 - Added Product Component w/ button buy and onAddToCart method
 - Implement Product Service
 - Added CartList component
+
+## 1.0.1
+
+### Added
+
+- Remove FirstComponent
+- Changed ProductModel
+- Add cart-item component
+- Implement cart service
+- Add subject for trigger CD in CartListComponent through ProductComponent
+- Convert Product and CartItem components in to "dumb" components
+- Add Material UI library